requirements.txt
fastapi==0.95.1
uvicorn[standard]==0.22.0
httpx==0.24.1
python-dotenv==1.0.0
tenacity==8.2.2
pydantic==1.10.11
pytest==7.4.0
pytest-asyncio==0.22.0
respx==0.20.0
black==24.3.0
mypy==1.3.0
